enable type checking for eslint.config.mjs again
there is conflict of type definitions between `@types/eslint` and
`typescript-eslint`. These packages provide their own types for
a config object in `eslint.config.mjs` and the types are not assignable.
`eslint-plugin-n` uses `@types/eslint` but `ts.config()` does not accept
the flat config object of the plugin.  It seems that the maintainer of
`typescript-eslint` won't improve this situation.

https://github.com/typescript-eslint/typescript-eslint/issues/8613#issuecomment-1983488262