prefer minimal implementation for HTTPS proxy
proxy-agent is unnecessarily large. Especially I don't like it depends
on quickjs (built with Emscripten) only for PAC proxy which we don't
need.
Instead I used proxy-from-env minimal package. This package is not
perfect because it refers `npm_config_*` environment variable for
unknown reasons however it is better than implementing the logic
manually.